Change keys in TimestampDB to checksums of the input string instead of storing the entire string in the DB.
This opens the door to partial build success tracking, which requires
storing a key that is the combination of the input and output filenames.
Representing the timestamps per graph edge will enable tracking that just
that edge has succeeded, thus preventing issues with the current system
where a partially successful build can prevent a dependency-only change
from building correctly on successive tries.